requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.3
oauthlib==3.1.0
Pillow==8.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.2
Werkzeug==1.0.1
wordcloud==1.8.1
